+++ src/base/kernel/interfaces/IConfig.h
@@ -78,6 +78,7 @@ public:
         PrintTimeKey         = 1007,
 
         // xmrig cpu
+        CPUKey               = 1024,
         AVKey                = 'v',
         CPUAffinityKey       = 1020,
         DryRunKey            = 5000,
@@ -99,6 +100,7 @@ public:
         OclMemChunkKey       = 1408,
         OclUnrollKey         = 1409,
         OclCompModeKey       = 1410,
+        OclKey               = 1411,
 
         // xmrig-proxy
         AccessLogFileKey     = 'A',