+++ src/core/config/Config_platform.h
@@ -45,6 +45,7 @@ static const char short_options[] = "a:c:kBp:Px:r:R:s:t:T:o:u:O:v:l:S";
 
 static const option options[] = {
     { "algo",                  1, nullptr, IConfig::AlgorithmKey          },
+#   ifdef XMRIG_FEATURE_HTTP
     { "api-worker-id",         1, nullptr, IConfig::ApiWorkerIdKey        },
     { "api-id",                1, nullptr, IConfig::ApiIdKey              },
     { "http-enabled",          0, nullptr, IConfig::HttpEnabledKey        },
@@ -52,6 +53,9 @@ static const option options[] = {
     { "http-access-token",     1, nullptr, IConfig::HttpAccessTokenKey    },
     { "http-port",             1, nullptr, IConfig::HttpPort              },
     { "http-no-restricted",    0, nullptr, IConfig::HttpRestrictedKey     },
+    { "daemon",                0, nullptr, IConfig::DaemonKey             },
+    { "daemon-poll-interval",  1, nullptr, IConfig::DaemonPollKey         },
+#   endif
     { "av",                    1, nullptr, IConfig::AVKey                 },
     { "background",            0, nullptr, IConfig::BackgroundKey         },
     { "config",                1, nullptr, IConfig::ConfigKey             },
@@ -76,13 +80,25 @@ static const option options[] = {
     { "user-agent",            1, nullptr, IConfig::UserAgentKey          },
     { "userpass",              1, nullptr, IConfig::UserpassKey           },
     { "rig-id",                1, nullptr, IConfig::RigIdKey              },
+    { "no-cpu",                0, nullptr, IConfig::CPUKey                },
+#   ifdef XMRIG_FEATURE_TLS
     { "tls",                   0, nullptr, IConfig::TlsKey                },
     { "tls-fingerprint",       1, nullptr, IConfig::FingerprintKey        },
+#   endif
+#   ifdef XMRIG_FEATURE_ASM
     { "asm",                   1, nullptr, IConfig::AssemblyKey           },
-    { "daemon",                0, nullptr, IConfig::DaemonKey             },
-    { "daemon-poll-interval",  1, nullptr, IConfig::DaemonPollKey         },
+#   endif
+#   ifdef XMRIG_ALGO_RANDOMX
     { "randomx-init",          1, nullptr, IConfig::RandomXInitKey        },
     { "randomx-no-numa",       0, nullptr, IConfig::RandomXNumaKey        },
+#   endif
+#   ifdef XMRIG_FEATURE_OPENCL
+    { "opencl",                0, nullptr, IConfig::OclKey                },
+    { "opencl-devices",        1, nullptr, IConfig::OclDevicesKey         },
+    { "opencl-platform",       1, nullptr, IConfig::OclPlatformKey        },
+    { "opencl-loader",         1, nullptr, IConfig::OclLoaderKey          },
+    { "opencl-no-cache",       0, nullptr, IConfig::OclCacheKey           },
+#   endif
     { nullptr,                 0, nullptr, 0 }
 };